requirements.txt
click>=6,<7
pros-cli
setuptools~=56.0.0